refactor: split recursion runtime owners
