+++ extensions/ypi/runtime.ts
@@ -8,6 +8,7 @@ export interface YpiRuntime {
 	systemPromptPath: string;
 	rlmQueryPath: string;
 	runtimeCorePath: string;
+	runtimeInternalDir: string;
 	cliAdapterPath: string;
 	legacyRlmQueryPath: string;
 }
@@ -25,6 +26,7 @@ export function resolveRuntime(importMetaUrl: string): YpiRuntime {
 		systemPromptPath: path.join(root, "SYSTEM_PROMPT.md"),
 		rlmQueryPath: path.join(root, "rlm_query"),
 		runtimeCorePath: path.join(root, "extensions", "ypi", "runtime-core.ts"),
+		runtimeInternalDir: path.join(root, "extensions", "ypi", "internal"),
 		cliAdapterPath: path.join(root, "extensions", "ypi", "cli.ts"),
 		legacyRlmQueryPath: path.join(root, "rlm_query.legacy"),
 	};
