+++ tests/test_pi_recursive_pack.sh
@@ -65,6 +65,7 @@ TARBALL_LIST="$(tar -tzf "$TARBALL")"
 assert_contains "tarball ships canonical extension" "package/extensions/recursive.ts" "$TARBALL_LIST"
 assert_contains "tarball ships native adapter" "package/extensions/ypi/native-tool.ts" "$TARBALL_LIST"
 assert_contains "tarball ships canonical runtime core" "package/extensions/ypi/runtime-core.ts" "$TARBALL_LIST"
+assert_contains "tarball ships internal child-process owner" "package/extensions/ypi/internal/child-process.ts" "$TARBALL_LIST"
 assert_contains "tarball ships retained native fallback" "package/extensions/ypi/legacy-native-tool.ts" "$TARBALL_LIST"
 assert_contains "tarball ships system prompt" "package/SYSTEM_PROMPT.md" "$TARBALL_LIST"
 assert_not_contains "tarball excludes the wrapper launcher" "package/ypi" "$TARBALL_LIST"
